Commit new changes (NOT WORKING)
- More organization with files and functions
- Code doesn't work anymore, need to fix function calls
