+++ Paradigmas/Projeto Debian/kernel.py
@@ -1,5 +1,5 @@
 import time
-from loading import Loading_animation
+from loading import LoadingAnimation
 
 class LinuxKernel:
     # Classe que representa o Kernel do Linux responsável pelo gerenciamento de hardware, recursos do sistema e interação com dispositivos.
@@ -7,24 +7,24 @@ class LinuxKernel:
         # Inicializa o Kernel do Linux com uma versão específica.
         self.version = version
 
-    def initialize(self):
+    def InitializeKernel(self):
         # Inicializa o Kernel do Linux.
         print("Inicializando o Kernel Linux", end='')
-        Loading_animation(1)
+        LoadingAnimation(1)
         print(f"Kernel Linux {self.version} inicializado com sucesso.")
         time.sleep(2)
 
-    def shutdown(self):
+    def ShutdownKernel(self):
         # Desliga o Kernel do Linux.
         print("Desligando o Kernel Linux", end='')
-        Loading_animation(1)
+        LoadingAnimation(1)
         print(f"Kernel Linux {self.version} desligado.")
         time.sleep(2)
 
 
 class LinuxKernelAPI:
     # Classe que fornece uma interface estável para os programas interagirem com o Kernel do Linux.
-    def interact_with_kernel(self): 
+    def InteractWithKernel(self): 
         print("Interagindo com o Kernel do Linux", end='')
-        Loading_animation(1)
+        LoadingAnimation(1)
         time.sleep(1)
