+++ Paradigmas/Projeto Debian/server.py
@@ -9,27 +9,18 @@ class Server:
         self.ip_address = ip_address
         self.machines = []
 
-    def initialize(self):
+    def InitializeServer(self):
         # Inicializa o servidor.
         print(f"Servidor {self.server_type} no endereço {self.ip_address} inicializado.")
         time.sleep(2)
 
-    def shutdown(self):
+    def ShutdownServer(self):
         # Desliga o servidor.
         print(f"Servidor {self.server_type} no endereço {self.ip_address} desligado.")
         time.sleep(2)
 
-    def add_machine(self, machine):
+    def AddMachine(self, machine):
         # Adiciona uma máquina ao servidor.
         self.machines.append(machine)
         print(f"Máquina adicionada ao servidor no endereço {self.ip_address}.")
         time.sleep(2)
-
-
-class Machine:
-    # Classe que representa uma máquina física ou virtual.
-    def __init__(self, machine_type, description):
-        # Inicializa uma máquina com um tipo e descrição específicos.
-        self.machine_type = machine_type
-        self.description = description
-        self.system_hardware = Hardware("Intel i7", "16GB", "1TB SSD")
